app/requirements.txt
appnope==0.1.0
backpack==0.1
blinker==1.4
bottle==0.12.13
certifi==2017.7.27.1
chardet==3.0.4
cleo==0.6.1
decorator==4.1.2
Faker==0.7.18
idna==2.6
inflection==0.3.1
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.10.2
lazy-object-proxy==1.3.1
orator==0.9.7
pastel==0.1.0
pendulum==1.2.5
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.1
ptyprocess==0.5.2
pyaml==16.12.2
Pygments==2.2.0
pylev==1.3.0
python-dateutil==2.6.1
pytz==2017.2
pytzdata==2017.2.2
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
traitlets==4.3.2
tzlocal==1.4
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
yml==0.0.1
